:green_heart: :recycle: Move type_test_runner.go into runners
