+++ parser/parsertest/runners/type_test_runner.go
@@ -1,4 +1,4 @@
-package parsertest
+package runners
 
 import (
 	"testing"
